Updated open and saveas
Updated JFileChooser so it opened on the currDirectory (determined by file currFile). If there is not currFile (and thus no currDirectory) then filechooser opened on default directory. This is partly done using a function that produces the JFileChooser (getJFileChooser())and it is used for both open and saveas so update open at the same time. This function also adds the filter. Added verification to savefilepath so that if it's null a file doesn't save. This also fixed a bug where files would save if save dialog cancelled.